refactor: move webhook features to own folder
