+++ source/php/DataProcessor/Handlers/Webhook/JsonDotHydrator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ModularityFrontendForm\Helper;
+namespace ModularityFrontendForm\DataProcessor\Handlers\Webhook;
 
 use Adbar\Dot;
