+++ source/php/DataProcessor/Handlers/Webhook/JsonDotHydratorTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ModularityFrontendForm\Helper;
+namespace ModularityFrontendForm\DataProcessor\Handlers\Webhook;
 
 use PHPUnit\Framework\TestCase;
